bugSlice.js file
Import createSlice from reduxjs toolkit. Create a slice const that equals a createSlice object that also expresses it self or returns a value because of the (). The createSlice has a name value pair of name with a bug value, initialState is the name with a [] empty array value and a reducers name with an object value that has a getBugs, createBugs, updateBug, markComplete which I think are methods