bugController.js
First import bugModel from Models folder which has bugModel. then export function called retrievbugs with a empty variable array called data. Add the data object with a method of push which passes a new bugModel object that express the name value pair of id with its number, name with its string value, details with its string value steps, version assigned creator, and time all of these names have string values assigned to them, priority has a number value of 3. A variable called sorted which equals a data object and the object is using it's sort method to sort a and b and then returns the result. sorted is returned.