userSlice.js file and some modification to bugSlice.js
First import createSlice from reduxjs toolkit. Add a slice const with a createSlice object that also express and returns a value, it has a name value pair of name and user, then initialState with an array object value and a reducers name with an object value which has a state that pushes two names. At the end we export slice.reducer as a default. Next add to bugSlice.js the two exports at the bottom.